execution: add an option to stream test output
This change adds a new option, "nocapture", that instructs execution
to stream test output to the standard output as soon as it arrives.
The option name is somewhat misleading as the output is captured anyway to
prevent concurrently executing tests to garble the output, but it's
still preferred for compatibility with the standard harness.

Fixes #1.